+++ src/app/shared/models/cart.models.ts
@@ -53,10 +53,9 @@ export interface Recipe extends List {
 export interface User {
   id: string;
   name: string;
-  surname: string;
   nickname: string;
   email: string;
-  password: string;
+  languaje: string;
 }
 
 export interface TokenUser {